+++ pacman.cpp
@@ -92,6 +92,7 @@ void Pacman::update(uint32_t time) {
     score += 10;
     pills_eaten++;
     pills_eaten_this_life++;
+    pill_eaten_time = time;
     // With a shortcut taken in one place, another piece of code gets more complicated. :D
     level_set(tile_pt, (flags & entityType::JUNCTION) ? 44 : (flags & entityType::CORNER) ? 49 : 0);
   }
@@ -101,6 +102,7 @@ void Pacman::update(uint32_t time) {
     score += 50;
     pills_eaten++;
     pills_eaten_this_life++;
+    pill_eaten_time = time;
     level_set(tile_pt, (flags & entityType::JUNCTION) ? 44 : (flags & entityType::CORNER) ? 49 : 0);
     start_power_timer();
     Ghosts::set_state(ghostState::FRIGHTENED);